skills/computer_use_requirements.txt
"""Dependencies for Computer Use Agent"""

# Install these packages to enable computer control:
# pip install -r skills/computer_use_requirements.txt

pyautogui>=0.9.54
Pillow>=10.0.0

# Optional but recommended:
# opencv-python>=4.8.0  # For advanced screen analysis
# numpy>=1.24.0         # For image processing
